refactor: move all shortcut related stuff
